cmake: Clear modules_list from the CACHE before setting it to empty value
The modules_list variable is updated with CACHE INTERNAL to append one module at a time when used. When using an IDE that automatically configures on save, this means that starting from a clean build directory (without cache), and wanting to check "WITH_GUI" to FALSE, will have the builds fail as the modules_list will contain gui targets, like g.gui.vdigit. The target ALL_MODULES would try to build g.gui.vdigit, but it doesn't appear anymore in the targets to build. It stays in the modules list, as it is stored in the CACHE.

The solution is to clear the cache variable, so using set(modules_list) defines a normal variable that is empty (and not having the cache variable previously hidden exposed)