man: Remove FileType argument types in parser standard options arguments (#5561)
* man: Write a help description for output argument of Parser standard option

* man: Remove usage of FileType argument type in parser standard options

The argparse FileType argument type will be deprecated in Python 3.14. It does not properly close files if a subsequent error in another FileType argument occurs when parsing. The suggestion is to use context managers after parsing the arguments. The implementation of the changes is inspired by a PR doing the same in the CPython repo.

* man: Remove proxies=None from urlopen, as it is for the old interface

* man: Add typing annotations to related changes of parser standard options

* man: Fix name of file by using outfile.name