CI(i18n): Create a workflow to update and extract strings to translate (#5752)
This PR adds a job in our periodic update workflow. For now, it runs once a month, plus can be triggered manually.

In order to better analyze the changes that are created with subsequent updates, I had to run some processing steps to the existing template and translation files with new flags, before running the step that actually collects new strings and line number locations. These file changes were remade on CI-only (not my machine), and with different configurations, to make sure nothing else was changing. The flags added/changed in the Makefile were done with consultation of Maris, the gettext family of tools docs, Weblate docs and guides, and other established projects translation setups.

The updating workflow and the PRs it creates are similar to what we use for keeping our config.sub and config.guess files updated. They have been tested successfully on my fork, in the situations with changes to do and no changes to apply. The logic uses the `--ignore-matching-lines` argument of git diff to ignore the lines that always change if the version-control modification time of the input files use change, which will always do in our case, even if no actual strings have changed. This limits useless PRs, and was tested in both cases.  Other established projects use a similar approach, with plain diff instead of git diff, in bash or perl scripts. This implementation is adapted to our reality.