r.slope.aspect: Increase precision of values compared in tests (#5796)
Follow-up from #5792.

This PR uses a smaller delta for the differences accepted between actual and reference values. The current tests were quite loose, and would probably not help us catch regressions unless the differences were huge. Since it is mostly a golden test, we want to know when anything changes.

* Increase r.slope.aspect precision

* Add more digits to the reference mean values of test_slope

* Increase r.slope.aspect precision to 1e-14, except where results fail

* Add note to use a smaller precision threshold or self.precision

* Apply suggestions from code review