CI(OSGeo4W): Upgrade image to windows-2022 (#5562)
The windows-2019 image is being removed in less than a month. We need to use windows-2022. The changes needed in our build_osgeo4w.sh script relate to the msys2 paths added to path. They need to use the location defined in the CI step, otherwise, another version already installed in the runner, but normally not on path, is used. On windows-2019, we didn’t see errors of this misconfiguration, but we did for windows-2022.

* CI(OSGeo4W): Upgrade image to windows-2022

* mswindows: Update path for dumpbin in mklibs.sh

* windows: Use MSYS2_LOCATION env var to create script from templates in build_osgeo4w.sh

* CI(OSGeo4W): Add MSYS2_LOCATION env var to build stage