doc: Generate zero, not type in Python examples
In the generated Python examples the type was leaking to the code, but does not make sense syntactically. For strings, we simply say string in quotes by default (if key description is not available), but for ints and floats, there is no such value (we don't have an example value), so this generates zero in the example. This will be likely wrong value for many cases, but it isvalid and also something the user will likely change (and it is easy to change).