docs: Adjust more code-blocks syntax and inside example sections in docstrings (#5862)
This is another round of making the formatting of code blocks either valid, better, or semantically better represented in example sections headers (that is, indented, but also represented inside the field list element in the HTML render, not just underneath).

One place, `Region.get_bbox()` (in python/grass/pygrass/gis/region.py), I created more docstring contents (for the return description and return type), to align with the rest around it.

* docs: Fix code block syntax in grass.exceptions

* docs: Fix code block formatting of grass.script.core

* docs: Fix code block formatting of grass.pygrass.gis.region

* docs: Make Example code blocks indented inside the header in the rendered page

* docs: Fix code block formatting in grass.script.core

* docs: Move up example of region_env() to group return and return type together

* docs: Make Example code blocks indented inside the header in the rendered page, by removing empty line between field list and code block

* docs: Make "Basic usage" code blocks indented inside the header in the rendered page, for grass.jupyter

* docs: Make "Usage" code blocks indented inside the header in the rendered pages

* docs: Make "Example" code blocks indented inside the header in the rendered pages, in grass.script and grass.temporal

* docs: Make "Basic usage" code blocks indented inside the header in the rendered pages, in grass.script

Also fixing the python code syntax when the code blocks were not being reformatted.