grass.script: Add more parameters to handle_errors
This add handler, stderr, and env to gs.handle_errors. This allows construction of better exception message and calling g.message with gs.fatal. Otherwise, it follows the same rules as error passed with the explicit kwargs dict parameter. Using stderr and handler to reduce confusion. errors is name of the type of handling in gs.handle_errors, but also the error messages in CalledModuleError.