g.mapset, g.mapsets: Specify stability of the outputs (#5847)
The current outputs are marked as plain and may possibly change in the future. For now, only with major release for backwards compatibility reasons, but in the future, any release can change the plain outputs. Implementing a separate, stable, shell-script-style output is a bigger endeavor, so the documentation now recommends JSON only.

This also changes the description of the tools to clarify what they do, and improves descriptions of the tests.