grass.script: Fallback to print when g.message fails (#5849)
A certain situation on Windows described in #5717 produces 'OSError: Cannot find the executable g.message' which is missing the original message. While such message is an indication of a deeper issue, the traceback misses the original message. This fallback approach show the message as the function was asked to do even when the subprocess had issues. It attaches the exception which may be confusing, but also may be quite helpful to understand the overall situation (which may not be related to the message itself).

When g.message is asked to print an error message with -e, it returns a non-zero return code. So, the error code really needs to be still ignored because printing an error results in a non-zero return code (likely to facilitate shell scripting).