grass.script: Pass env to debug_level
To get a debug level Python code calls debug_level function which in turn needs to call g.gisenv because the debug level is stored in GRASS gisenv variable (and not as a system environment variable). For call to g.gisenv, one needs to have a session (GISRC) set up, so env is needed when it is only a local one (and not os.environ).

Experimentally, this also removes the monkeypatch fixture from the recently added test for g.list which, without the monkeypatch, failed on Windows with GISRC missing from g.gisenv indicating that os.environ was set up with the runtime (PATH etc.), but didn't have an active session (GISRC etc.). This should pass the correct local environment through env all the way to g.gisenv. (Additionally, the fixture scope is changed to module, so that it is created only once.)