grass.script: Do not change dir to create XY project (#5918)
Changing the current working directory to create the files might have been inherited from a shell workflow, but it is not needed to create the files, and it just complicates the code. Using pathlib for everything, not a mix of os.mkdir and Path.mkdir. We don't create parents, but only the project and mapset directories.