grass.script: More carefully setup env in create_project (#5854)
Newly, the env in create_project is not set up not only when GISBASE is not set, but also when GISBASE is not in PATH which at this point indicates that tools will not run.

The env setup is done only when needed, so simple XY does not trigger creation of the environment, and so it is more lightweight.

The code to ensure runtime environment is moved to grass.script.setup, broken down and made reusable. The difference for usage in create_project is that os.environ is now always copied even if it already has GISBASE (we saved that copy operation before if GISBASE was already present). The handling of env, however, is consistent with other functions in grass.script.setup.

With faster create_project, I replaced the old gs.core._create_location_xy calls by simple gs.project calls in tests. Removed pylint: disable=protected-access, and removed one variable (constructing the path directly from a string).

A new test is checking the case when overwriting a project and creating an XY one which (for XY project) should also test cleaning GISRC without processing g.proj run.