i18n: Replace strings constructed by parts with complete strings (#5938)
- Added translation comments for context on the units
- Made strings reorderable for translation
A few strings were flagged by when running xgettext, mentioning that format string with unnamed arguments cannot be properly localized:
The translator cannot reorder the arguments.
Please consider using a format string with named arguments,
and a mapping instead of a tuple for the arguments.

- Also shoved in fixes of some typos encountered during that search.


* i18n: Replace strings constructed by parts with complete strings

* i18n: Add translation comments to units

* i18n: Make strings reorderable for translation

A few strings were flagged by when running xgettext, mentioning that format string with unnamed arguments cannot be properly localized:
The translator cannot reorder the arguments.
Please consider using a format string with named arguments,
and a mapping instead of a tuple for the arguments.

* docs: Fix various typos