grass.script: Add env, handler, stderr to handle_errors (#5875)
This add handler, stderr, and env to grass.script.handle_errors. This allows construction of better exception message and calling g.message with grass.script.fatal when global environment is not used. Otherwise, it follows the same rules as error passed with the explicit kwargs dict parameter. Using stderr and handler to reduce confusion with errors. The term errors is name of the type of handling in grass.script.handle_errors, but also the error messages in CalledModuleError.

It also reads env from kwargs which will make grass.script.fatal with g.message work with run_command and others which call the function without knowing all the parameters passed by the caller.

---------

Co-authored-by: Edouard Choinière <27212526+echoix@users.noreply.github.com>