grass.script: Allow a leading dash in flags parameter (#5952)
Instead of raising a ScriptError, accept a leading dash for the flags parameter in run_command family of functions. While no dash is the best practice in Python, and therefore no examples or documentation are changed, leading dash is required in the command line and appears in documentation and messages, so allowing it in Python makes it little easier on the user.

In command line, leading dash is required for a group a flag letters which would be the most typical use of multiple flags. Since this requirement is not going away - it is part of the command line syntax itself, we ease the transition between the two environments by supporting the leading dash in Python. Dashes inside the flags string, including space-separated flag characters with dashes, are not supported because that would basically mean processing the space separated command line syntax (trivial to process in this case, but does go too far from Python list of flag letters into full command line syntax). Even with a leading dash, the code looks sufficiently like Python.

In documentation and messages (typically warnings or errors), we use mix of standalone character (possibly in single or double quotes) and a dash followed by a character. While it may be worth polishing this across the documentation and code, and perhaps going with characters only (because they apply in both command line and Python), the command line examples will always have dashes, so supporting the dash for copy-pasting scenarios would be nice. (The new generated Python documentation for tools does not include the dashes, so it supports more doing the right thing - characters only, but command line examples are currently prevalent and will always be preset.)