grass.script: Remove PATH modification, fix GUI scripts in get_commands
get_commands modifies the system PATH environment variable in the global environment as a side effect unrelated to what it is actually doing. This is not needed to generate the return value and it is not expected outcome from running the get_commands function. This simply removes the line which is modifying PATH and adds test code for it.

The block for the GUI scripts where the PATH modification was didn't work because the path to GUI scripts was wrong. Also, it didn't add the tools in the same way as for the other directories on Windows. This fixes the path, uses the same scan procedure, and tests for it in the tests. (Assuming that adding these GUI scripts is actually the desired behavior when calling get_commands which may or may not be the case, but it was the intention of the original code.)