grass.script: Remove collection GUI scripts from get_commands (#5971)
Completely removes GUI scripts from get_commands because they are too special. All the tools get_commands gives you should be on PATH and behave the same. We may consider moving these special tools to scripts.

The block for the GUI scripts didn't work because the path to GUI scripts was wrong. Also, it didn't add the tools in the same way as for the other directories on Windows.

The original get_commands modified the system PATH environment variable in the global environment as a side effect unrelated to what it is actually doing. This is not needed to generate the return value and it is not expected outcome from running the get_commands function. A tests now tests that this not reintroduced.

Renames the test to xxx_test to enable automatic testing.

Fixes the Windows Python script part of the test (extension includes a dot).