grass.tools: Add Tools API to generated tools doc
Concept of how this could look like. The formatting needs more work. The parameters are the same as grass.script except io.StringIO which will be difficult to generate with the current metadata because the parser does not know about stdin in this sense (the validity is resolved while opening the file by a library function call at tool-level, but during command line parsing, the parser merely allows for dash being provided instead of a filename).