build: refactor grass.py in preparation for FHS (#5933)
The transitional period from the traditional GRASS installation
structure to a Filesystem Hierarchy Standard (FHS) complying structure
(enabled by CMake build) requires an alternative way to find resources.
Currently all resources are located in relation to GISBASE. That is no
longer the case with a FHS installation.

In addition, and instead of GISBASE, a number of environment
variables will be needed. Setting up all these variables would bloat
the grass.py file, therefore this update moves this to the grass.app
Python module.

Summary of changes:

- python/grass/app/resource_paths.py is a configurable file, where 
  default installation paths to resources are configured as well as
  other build time properties
- to simplify Conda adoption, the variable GRASS_PREFIX is added, which
  is essentially equal to installation prefix, but easily changeable
  if needed.
- grass.py need to find the Python module to be able to set necessary
  variables, a new variable GRASS_PYDIR is configured into the file to
  do just this.
- update copy_python_files_in_subdir.cmake to be able to exclude files