CQ: Replace crlf check with validation against an .editorconfig file (#6070)
Simplified and adapted from  #4792.
Fixes https://github.com/OSGeo/grass/pull/6066#issuecomment-3075510813

The action that checks for CRLF has been failing since at least this weekend.
Considering that that action has low chances of beeing fixed in a timely matter, I replaced the functionality with an .editorconfig file and validating against it with editorconfig-checker. Compared to #4792, here I don’t attempt to make the line ending changes persist in the repo through .gitattributes, and ignore a bit more of non-ready config sections (since the functionality replaced was less detailed than what the .editorconfig I already had almost working).


I made the pre-commit run always, and made it faster in consequence to not add a big impact. With hot cache, it is taking 1min-1min10, which is about the same. I knew pre-commit could be faster in CI as I already saw how fast it was in ruff’s own CI when submitting one of my PRs there once. It can still be a bit faster, but it’s enough for today.

* CQ: Add initial .editorconfig file with basic settings

* CI: Add EditorConfig validation to super-linter workflow

* CI: Remove CRLF check from additional_checks.yml workflow

* CQ: Add editorconfig-checker to pre-commit config

* CI: Always run pre-commit in CI

* Update .editorconfig with new whitespace settings for mswindows

* CI: Run pre-commit step with uvx

* CI: Add caching for pre-commit checks in workflow

* Update .editorconfig for Windows batch files

* CI: Remove paths-filter step from additional_checks.yml

* Update additional_checks.yml step setup-uv