lib/parser: Use the first allowed values in example (#6079)
In the generated example for each tool, use the first allowed value from the list of allowed values in the options field. This can be done for strings only because numbers may have ranges in the options field. Before, simply the type, i.e., string would be printed unless key_desc was defined. For r.series, this is replacing 'string' by 'average' as a value for method in the example.

Apply the same to Python and CLI and sync CLI to Python which adds the fallback number values to CLI example.