r.mapcalc: remove static variables from mode and median to fix parallelization (#6087)
This gets rid of static variables from the r.mapcalc functions for mode and median. The static variables were there to not allocate for every row, but it complicates parallelization since that can't generally be done in parallel with static variables, see #5742. Since the size of the array is typically very small (the number of arguments of the median() or mode() function, can be raster or number), it will use array on stack and allocate only for larger sizes.

It may be slower due to more stack or heap needed, but it might be faster for cases within the threshold (using heap allocated memory instead of realloc). In any case, it will reliably work with parallelization. General solution would be using a context, but that would require a large change.