lib/parser: Use None as default for overwrite (#6076)
Use None, not False, as a default value for overwrite, verbose etc. None is more appropriate because the default behavior is influenced by the environment and, very generally, by how the library is set up. While False is correct in terms of the corresponding command line not being set, the False value may create a false perception that overwriting will be disabled, but in fact overwriting will be driven by the environment.